+++ ndlib/SIRModel.py
@@ -41,5 +41,4 @@ class SIRModel(DiffusionModel):
         self.status = actual_status
         self.actual_iteration += 1
 
-        # return self.actual_iteration, actual_status
-        return self.actual_iteration, delta
+        return self.actual_iteration - 1, delta