+++ ndlib/SISModel.py
@@ -41,8 +41,7 @@ class SISModel(DiffusionModel):
         self.status = actual_status
         self.actual_iteration += 1
 
-        # return self.actual_iteration, actual_status
-        return self.actual_iteration, delta
+        return self.actual_iteration - 1, delta
 
     def complete_run(self, max_iteration=200):
         system_status = []